backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
elevenlabs==0.2.27
requests==2.31.0
gunicorn==21.2.0

